Fix issue #394 (#395)
* Add xfail test for issue 394

* Fixes for #394

Modify the JINJA_FUNCTION_ADD_CONCAT regular expression to force a space before the +
We could have a real concat expression without space before the +, but I haven't seen it in recipes.

This is to allow version number like 1.2.1.dev1+g3df5418 to be interpreted as string and not expression.